move Quantizer class to base.py, set __all__
